// apps/api/generated/prisma/models/Theme.ts
export type ThemeModel = runtime.Types.Result.DefaultSelection<Prisma.$ThemePayload>
// ...
export type $ThemePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Theme"
  objects: {
    presentations: Prisma.$PresentationPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    displayName: string
    description: string
    isBuiltIn: boolean
    primaryColor: string
    secondaryColor: string
    accentColor: string
    backgroundColor: string
    textColor: string
    headingFont: string
    bodyFont: string
    colorPalette: runtime.JsonValue
    createdAt: Date
  }, ExtArgs["result"]["theme"]>
  composites: {}
}